requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
openai==1.12.0
aiohttp==3.9.3
Pillow==10.2.0
gunicorn==21.2.0